fix: fix some conversations
