+++ .gitignore
@@ -2,6 +2,6 @@ client/node_modules
 client/build
 client/.env
 server/dist
+server/config/local.json
 server/node_modules
-server/config/default.json
 server/docker-compose.yml
