--- server/config/default.example.json
@@ -1,10 +1,0 @@
-{
-  "PORT": 0,
-  "DB_NAME": "",
-  "DB_USER": "",
-  "DB_PASSWORD": "",
-  "DB_HOST": "",
-  "DB_PORT": 0,
-  "PUBLIC_URL": "",
-  "BASE_URL": ""
-}
