+++ server/config/local.example
@@ -1,0 +1,3 @@
+{
+  "DB_PASSWORD": ""
+}
