+++ server/docker-compose.example
@@ -1,20 +1,5 @@
 version: "3.8"
 services:
-  url_backend:
-    container_name: url_backend
-    image: francescoxx/nsp-template:0.0.2
-    build:
-      context: .
-    ports:
-      - "5000:5000"
-    environment:
-      - EXTERNAL_PORT=
-      - DB_USER=
-      - DB_PASSWORD=
-      - DB_NAME=
-      - DB_HOST=
-    depends_on:
-      - url_backend_db
   url_backend_db:
     container_name: url_backend_db
     image: "postgres:12"
