+++ server/package.json
@@ -50,6 +50,7 @@
     "geoip-lite": "^1.4.7",
     "http-status-codes": "^2.2.0",
     "ip": "^1.1.8",
+    "nanoid": "^3.3.6",
     "pg": "^8.10.0",
     "pg-hstore": "^2.3.4",
     "sequelize": "^6.29.3",
